pyproject.toml
[project]
name = "mdl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Clemens Koza", email = "ckoza@tgm.ac.at" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "moodlepy>=0.24.1",
    "requests>=2.32.3",
    "ruamel-yaml>=0.18.14",
    "typer>=0.16.0",
    "typst>=0.13.3",
]

[project.scripts]
mdl = "mdl.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
moodlepy = { git = "https://github.com/SillyFreak/moodlepy", branch = "improvements" }

[tool.hatch.build.targets.wheel]
packages = ["src/mdl"]
